fix(hover): separated logic for root node and extend
This caused issues in case project uses same naming when extending jobs.
So base job has name "test" and the job extending it has name "test".
This caused flakyness because order is not guaranteed in maps. So in
some cases current location was resolved to base job instead of hovered
job.